Fix unsharp app

+++ apps/hardware_benchmarks/apps/unsharp/unsharp_generator.cpp
@@ -48,7 +48,7 @@ public:
 
         // create a grayscale image
         Func gray;
-        gray(x, y) = cast<uint8_t>((77 * cast<uint16_t>(hw_input(x, y, 0))
+        gray(x, y) = cast<uint16_t>((77 * cast<uint16_t>(hw_input(x, y, 0))
                                     + 150 * cast<uint16_t>(hw_input(x, y, 1))
                                     + 29 * cast<uint16_t>(hw_input(x, y, 2))) >> 8);
         
@@ -61,11 +61,11 @@ public:
 
         // sharpen the image by subtracting the blurred image
         Func sharpen;
-        sharpen(x, y) = cast<uint8_t>(clamp(2 * cast<uint16_t>(gray(x, y)) - blur(x, y), 0, 255));
+        sharpen(x, y) = cast<uint16_t>(clamp(2 * cast<uint16_t>(gray(x, y)) - blur(x, y), 0, 255));
 
         // find the ratio of sharpened and original image
         Func ratio;
-        ratio(x, y) = cast<uint8_t>(clamp(cast<uint16_t>(sharpen(x, y)) * 32 / max(gray(x, y), 1), 0, 255));
+        ratio(x, y) = cast<uint16_t>(clamp(cast<uint16_t>(sharpen(x, y)) * 32 / max(gray(x, y), 1), 0, 255));
 
         // Use the ratio to sharpen the input image.
         Func hw_output;
